libaxl: * [fix] Fixed wrong int type declaration passed in into   gettenv_s in windows which causes possible stack corruption   under 64bits plaforms.

+++ trunk/libaxl/src/axl_log.c
@@ -64,7 +64,7 @@ axl_bool     debug_color_enabled = axl_false;
 axl_bool      axl_log_is_enabled (void) 
 {
 #if defined(AXL_OS_WIN32) && ! defined(__GNUC__)
-	int requiredSize;
+	size_t requiredSize;
 #endif
 
 	if (not_executed) {
@@ -90,7 +90,7 @@ axl_bool      axl_log_is_enabled (void)
 axl_bool    axl_log_color_is_enabled (void)
 {
 #if defined(AXL_OS_WIN32) && ! defined(__GNUC__)
-	int requiredSize;
+	size_t requiredSize;
 #endif
 
 	if (not_executed_color) {
